Overhaul queue generation to actually utilize the queue names from the collector.
Includes first draft of the command-line version of the collector.
