+++ src/collector_to_agis
@@ -1,0 +1,28 @@
+#!/usr/bin/python
+
+import os
+import optparse
+
+if 'CONDOR_CONFIG' not in os.environ:
+    os.environ['CONDOR_CONFIG'] = '/etc/condor-ce/condor_config'
+import htcondor
+
+import htcondorce.agis_compat
+
+def main():
+    parser = optparse.OptionParser()
+    parser.add_option("-p", "--pool", dest="pool", default="collector.opensciencegrid.org")
+    parser.add_option("-v", "--verbose", dest="verbose", default=False, action="store_true")
+    opts, args = parser.parse_args()
+
+    if opts.verbose:
+        if 'TOOL_DEBUG' not in htcondor.param:
+            htcondor.param['TOOL_DEBUG'] = 'D_FULLDEBUG'
+        htcondor.enable_debug()
+
+    htcondorce.agis_compat.agis_compat_main(pool=opts.pool)
+
+
+if __name__ == '__main__':
+    main()
+
